data_classification/frame_classification/requirements.txt
# Streamlit
streamlit>=1.28.0

# Datenbank
psycopg2-binary>=2.9.0
duckdb>=0.10.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Visualization
plotly>=5.15.0

# NLP
spacy>=3.7.0
nltk>=3.8.0

# Environment
python-dotenv>=1.0.0
